backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
click==8.1.3
Django==4.0.4
django-cors-headers==3.12.0
django-simple-chatbot==0.0.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
joblib==1.1.0
nltk==3.7
PyJWT==2.4.0
pytz==2022.1
regex==2022.4.24
sqlparse==0.4.2
tqdm==4.64.0
